Add power type panel and show tooltips on influence graph
When the deck list changes, update the count of card type tags
for all elements in the document with class "power-type-count".
The "tag" attribute is used to specify which tag to count.

Draw dots along the curves in the influence graph at the position
where there is a card with power and influence corresponding
to that influence color and turn number.

When mousing over influence dots on the graph, show a tooltip
listing an exact percentage and cards relevant to that point
on the graph.

When hovering over the cost cell in the power odds table,
show a tooltip with the names of the cards which have that
cost.

In the power odds table, clarify the power requirement in the tooltip
by presenting the text as "3 Power - 82%" rather than as "3 - 82%".

Switched back from labeling the graph and table in terms of number
of draws to number of turns, with the assumption that there is one
draw per turn.  This makes the card power costs align better
with their position in the graph and column in the table.